refactor(ui): native sdk util
native sdk util
